@@ -3,53 +3,71 @@
   "version": "1.0.0",
   "description": "The frontend application for projects",
   "scripts": {
-    "start": "webpack-dev-server",
-    "dist": "webpack --config ./webpack-prod.config.js",
-    "lint": "eslint src",
-    "test": "NODE_ENV=test mocha --recursive --compilers js:babel-core/register",
-    "test:watch": "npm test -- --watch"
+    "start": "node scripts/start.js",
+    "build": "node scripts/build.js",
+    "test": "node scripts/test.js --env=jsdom"
   },
   "author": "",
   "license": "UNLICENSED",
   "devDependencies": {
-    "babel-core": "^6.8.0",
-    "babel-eslint": "^6.0.4",
-    "babel-loader": "^6.2.4",
-    "babel-preset-airbnb": "^2.0.0",
-    "babel-preset-es2015": "^6.6.0",
-    "babel-preset-react": "^6.5.0",
-    "babel-preset-stage-1": "^6.5.0",
-    "css-loader": "^0.23.1",
-    "enzyme": "^2.0.0",
-    "eslint": "^2.9.0",
-    "eslint-config-airbnb": "^9.0.1",
-    "eslint-loader": "^1.3.0",
-    "eslint-plugin-import": "^1.7.0",
-    "eslint-plugin-jsx-a11y": "^1.2.0",
-    "eslint-plugin-react": "^5.0.1",
-    "expect": "^1.20.1",
-    "fetch-mock": "^4.6.0",
-    "jsdom": "^8.1.0",
-    "less": "^2.6.1",
-    "less-loader": "^2.2.2",
-    "mocha": "^2.5.3",
+    "autoprefixer": "^9.2.1",
+    "babel-eslint": "^10.0.1",
+    "babel-loader": "^7.1.2",
+    "babel-preset-react-app": "^3.1.1",
+    "case-sensitive-paths-webpack-plugin": "^2.1.2",
+    "css-loader": "^1.0.0",
+    "enzyme": "^3.7.0",
+    "eslint": "^5.7.0",
+    "eslint-config-airbnb": "^17.1.0",
+    "eslint-loader": "^2.1.1",
+    "eslint-plugin-import": "^2.14.0",
+    "eslint-plugin-jsx-a11y": "^6.1.2",
+    "eslint-plugin-react": "^7.11.1",
+    "expect": "^23.6.0",
+    "fetch-mock": "^7.1.0",
+    "file-loader": "^2.0.0",
+    "html-webpack-plugin": "2.29.0",
+    "jsdom": "^12.2.0",
+    "less": "^3.8.1",
+    "less-loader": "^4.1.0",
+    "mocha": "^5.2.0",
+    "postcss-loader": "^3.0.0",
     "react-addons-test-utils": "^15.1.0",
-    "react-hot-loader": "^1.3.0",
+    "react-dev-utils": "^5.0.1",
+    "react-hot-loader": "^4.3.11",
     "redux-mock-store": "^1.0.3",
-    "style-loader": "^0.13.0",
-    "webpack": "^1.13.0",
-    "webpack-dev-server": "^1.14.1"
+    "source-map-loader": "^0.2.1",
+    "style-loader": "0.19.0",
+    "tsconfig-paths-webpack-plugin": "^3.2.0",
+    "url-loader": "^1.1.2"
   },
   "dependencies": {
+    "@babel/core": "^7.0.0",
+    "@babel/preset-react": "^7.0.0",
+    "babel-core": "^6.26.3",
+    "chalk": "^2.4.1",
+    "extract-text-webpack-plugin": "^3.0.2",
+    "fs-extra": "^7.0.0",
+    "history": "^4.7.2",
     "immutable": "^3.8.1",
-    "react": "^15.0.2",
-    "react-dom": "^15.0.2",
-    "react-redux": "^4.4.5",
-    "react-router": "^2.4.0",
-    "material-ui": "^0.15.1",
-    "react-tap-event-plugin": "^1.0.0",
-    "redux": "^3.5.2",
-    "reselect": "^2.5.1",
-    "whatwg-fetch": "^1.0.0"
+    "material-ui": "^0.20.2",
+    "prop-types": "^15.6.2",
+    "react": "^16.4.2",
+    "react-dom": "^16.4.2",
+    "react-redux": "^5.0.7",
+    "react-router": "^3.2.1",
+    "react-tap-event-plugin": "^3.0.3",
+    "redux": "^4.0.1",
+    "reselect": "^4.0.0",
+    "sw-precache-webpack-plugin": "^0.11.5",
+    "webpack": "3.8.1",
+    "webpack-dev-server": "2.9.4",
+    "webpack-manifest-plugin": "^2.0.4",
+    "whatwg-fetch": "^3.0.0"
+  },
+  "babel": {
+    "presets": [
+      "react-app"
+    ]
   }
 }
