@@ -1,7 +1,8 @@
-import React from 'react';
-import { browserHistory } from 'react-router';
+import * as React from 'react';
+import createHistory from 'history/createBrowserHistory';
+import PropTypes from 'prop-types';
 import ProjectRow from './projectrow';
-
+import { browserHistory } from 'react-router';
 
 const ProjectsList = (props) => {
   if (props.projects.loading === true) {
@@ -19,18 +20,20 @@ const ProjectsList = (props) => {
         <button
           onClick={() => browserHistory.push('/projects/new')}
           id='add-employee-button'
-          className='mdl-button mdl-js-button mdl-button--fab mdl-button--mini-fab'
+          className='mdl-button mdl-js-button mdl-button--fab mdl-button--mini-fab new-project-button'
         >
           <i className='material-icons dark-gray'>add</i>
         </button>
       </div>
-      {projectsRows}
+      <div className="projectRows">
+        {projectsRows}
+      </div>
     </div>
   );
 };
 
 ProjectsList.propTypes = {
-  projects: React.PropTypes.object.isRequired
+  projects: PropTypes.object.isRequired
 };
 
 export default ProjectsList;
