@@ -1,20 +1,17 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import injectTapEventPlugin from 'react-tap-event-plugin';
 import { Provider } from 'react-redux';
 import { compose, createStore, applyMiddleware } from 'redux';
-import apiMiddleware from './middleware/api';
 import { Router, browserHistory } from 'react-router';
 import getMuiTheme from 'material-ui/styles/getMuiTheme';
 import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
-
+import apiMiddleware from './middleware/api';
 import routes from './routes';
 import reducers from './reducers';
 
-require('../styles/main.less');
+import './styles/main.less';
 
 // needed for material-ui, for now
-injectTapEventPlugin();
 
 const createStoreWithMiddleware = compose(
     applyMiddleware(apiMiddleware),
